requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-cloud-aiplatform==1.75.0
google-auth==2.37.0
google-cloud-speech==2.28.0
requests==2.32.3
